added comments for each regex, removed the file code from get_bob_ross.py
