+++ after_17.py
@@ -5,7 +5,7 @@ import help_functions
 
 def message_after_hour(bob_ross_dict, startHour):
     counter = 0
-    hour_regex = 'T(.*?):' # finder timerne i hver bob_ross key
+    hour_regex = 'T(.*?):' # finds the hours in each bob ross key: the string between the first capitalT and the first colon
     hour_dict = help_functions.line_pattern(bob_ross_dict, hour_regex)  
     for key in hour_dict:
         if int(key) >= startHour:
