+++ get_bob_ross.py
@@ -15,48 +15,16 @@ import after_17
 import count_lines
 import count_user_names
 import help_functions
-
-
-def download(from_url, to_file):
-    if not os.path.isfile(to_file):
-        req.urlretrieve(from_url, to_file)
+import ruined_word
 
 
 if __name__ == '__main__':
-    try:
-        global file_name
-        _, url, file_name = sys.argv
-    except:
-        try:
-            _, url = sys.argv
-            file_name = os.path.basename(url)
-        except:
-            try:
+    bob_ross_file = help_functions.bob_ross_file()
 
-                # open file
-                cfg_file = 'list_of_files.txt'
-                with open(cfg_file) as fp:
-                    for line in fp:
+bob_ross_dict = help_functions.convert_file_dict(bob_ross_file)
 
-                        file_name = os.path.basename(line.rstrip())
-                        url = line
-                        download(url, file_name)
-                sys.exit(0)
-            except Exception as e:
-                print(__doc__)
-                sys.exit(1)
-    download(url, file_name)
-
-
-def convert_file_dict():
-    with open(file_name, encoding='utf8') as fp: 
-        global bob_ross_dict
-        bob_ross_dict = {key: value for key, value in [
-            line.strip().split(None, 1) for line in fp]}
-
-convert_file_dict()
 print(count_lines.countLines(bob_ross_dict))
 print(count_user_names.get_username_count(bob_ross_dict))
 print(after_17.message_after_hour(bob_ross_dict, 17))
 print(most_used_words.count_frequency_words(bob_ross_dict))
-print(len(help_functions.line_pattern(bob_ross_dict, ", [\'\"](.*?): ")))
+print(ruined_word.ruin_work())
